support for cumulative progress statements
This is useful when the order that the steps involved in running a job is non-deterministic.

For instance, instead of stating that step A runs from 0-10% and
step B runs from 10-100%, you can instead say that step A runs
for 10% of the total time and B runs for 90% of the total time.

- To use a range, use two integers with a dash between: DD-DD
- To add a percentage, use a single integer with no dash: DD

...where DD is an integer between 0 and 100.